Work specifics
Apparently the copilot-server used in later versions of copilot.el
does not work with the company github profile along with personal
profile.   Completions are missing and I think the problem is
that the server needs to be logged in as both company and personal?

Not sure at all.   But this version of copilot.el works inside company.

Need more debugging.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>